main - create, findAll e findById ja estao executando com sucesso

+++ src/main/java/com/itau/ms_desafio_itau/repository/AutomovelRepository.java
@@ -3,6 +3,6 @@ package com.itau.ms_desafio_itau.repository;
 import com.itau.ms_desafio_itau.entities.Automovel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutomovelRepository extends JpaRepository<Automovel, Integer> {
+public interface AutomovelRepository extends JpaRepository<Automovel, Long> {
 
 }
